Update loadrube.js
When applying properties to bodies, fixtures, and joints, there is lots of repetitious code to first see if a property is present in the source object, and then copy its contents over to the newly-instantiated object.  Added the "applyProperties" function at the top and DRY-ed things up.

Plus there's a switch statement now for testing the body type, instead of three "if" statements.